car_repair_backend/requirements.txt
annotated-types==0.7.0
asgiref==3.11.0
beautifulsoup4==4.14.3
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
contextlib2==21.6.0
cryptography==46.0.3
dj-database-url==3.0.1
Django==4.2.27
django-cors-headers==4.9.0
django-ninja==1.5.1
django-ninja-extra==0.30.8
django-ninja-jwt==5.4.3
gunicorn==23.0.0
idna==3.11
injector==0.24.0
lxml==6.0.2
packaging==25.0
psycopg2-binary==2.9.11
pycparser==2.23
pydantic==2.12.5
pydantic-settings==2.11.0
pydantic_core==2.41.5
PyJWT==2.10.1
python-dotenv==1.2.1
requests==2.32.5
soupsieve==2.8.3
sqlparse==0.5.5
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
whitenoise==6.11.0
